docs/requirements.txt
furo
jinja2<3.1
nbsphinx>=0.8.0
nb-black==1.0.7
matplotlib==3.3.3
sphinx-copybutton
networkx>=2.0.0
pandoc
sphinxcontrib.bibtex
jupytext
ipython
ipykernel
watermark
sphinxext-opengraph
blackjax
dm-haiku
ipywidgets
# Install GPJax istself
.